- resolve session comments
