+++ datasource/remote/build.gradle
@@ -8,6 +8,14 @@ android {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            buildConfigField 'String', "ApiKey", AnotherMovieApp_ApiKey
+            resValue 'string', "api_key", AnotherMovieApp_ApiKey
+        }
+        debug {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            buildConfigField 'String', "ApiKey", AnotherMovieApp_ApiKey
+            resValue 'string', "api_key", AnotherMovieApp_ApiKey
         }
     }
 }
