+++ datasource/remote/src/main/AndroidManifest.xml
@@ -1,4 +1,5 @@
-<manifest package="com.egdroid.datasource.remote">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.egdroid.datasource.remote">
 
-    <uses-permission name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 </manifest>
