+++ datasource/remote/src/main/java/com/egdroid/datasource/remote/movie/Extensions.kt
@@ -1,0 +1,13 @@
+package com.egdroid.datasource.remote.movie
+
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkInfo
+
+fun isNetworkConnected(context: Context): Boolean {
+    val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE)
+    return if (connectivityManager is ConnectivityManager) {
+        val networkInfo: NetworkInfo? = connectivityManager.activeNetworkInfo
+        networkInfo?.isConnected ?: false
+    } else false
+}
