+++ datasource/remote/src/main/java/com/egdroid/datasource/remote/movie/PopularMovieService.kt
@@ -7,6 +7,6 @@ import retrofit2.http.GET
 
 interface PopularMovieService {
 
-    @GET("popular?api_key=828157095a54ed8f68b8882624ed7660&language=en-US&page=1")
+    @GET("popular")
     fun getMovies(): Call<MovieResponse>
 }
