--- datasource/remote/src/main/java/com/egdroid/datasource/remote/utils/NetworkUtils.kt
@@ -1,15 +1,0 @@
-package com.egdroid.datasource.remote.utils
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkInfo
-
-object NetworkUtils {
-    fun isNetworkConnected(context: Context): Boolean {
-        val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE)
-        return if (connectivityManager is ConnectivityManager) {
-            val networkInfo: NetworkInfo? = connectivityManager.activeNetworkInfo
-            networkInfo?.isConnected ?: false
-        } else false
-    }
-}
